% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_basic.R
\name{auth_basic}
\alias{auth_basic}
\title{Basic authentication plugin}
\usage{
auth_basic(
  authenticator,
  get = NULL,
  head = NULL,
  post = NULL,
  put = NULL,
  delete = NULL,
  connect = NULL,
  options = NULL,
  trace = NULL,
  patch = NULL,
  all = NULL,
  name = "BasicAuth",
  realm = "private",
  ignore_trailing_slash = FALSE
)
}
\arguments{
\item{authenticator}{A function that takes a username and password and
returns \code{TRUE} if the pair is valid, and \code{FALSE} otherwise}

\item{get, head, post, put, delete, connect, options, trace, patch, all}{Character
vectors with paths to add authentication to up front. Additional paths can be
added afterwards using the \code{add_handler()} method.}

\item{name}{The name of the plugin}

\item{realm}{The realm this authentication corresponds to. Will be returned
to the client on a failed authentication attempt to inform them of the
credentials required though most often these days it is kept from the user.}

\item{ignore_trailing_slash}{Logical. Should the trailing slash of a path
be ignored when determining if a request should be authenticated. Setting
this will not change the request or the path associated with it but just ensure that
both \code{path/to/resource} and \verb{path/to/resource/} ends up being authenticated}
}
\value{
A \code{AuthBasicRoute} object
}
\description{
Basic authentication is a HTTP scheme that sends username and password as a
\code{:} separated, base64 encoded string in the authorization header. Because it
is effectively send in plain text (base64 encoding can easily be decoded)
this should only ever be used along with other security measures such as
https/ssl to avoid username and passwords being snooped from the request.
}
