% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireproof-package.R
\docType{package}
\name{fireproof-package}
\alias{fireproof}
\alias{fireproof-package}
\title{fireproof: Authentication and Authorization for 'fiery' Servers}
\description{
Provides a plugin for 'fiery' that supports various forms of authorization and authentication schemes. Schemes can be required in various combinations or by themselves and can be combined with scopes to provide fine-grained access control to the server.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/fireproof}
  \item \url{https://thomasp85.github.io/fireproof/}
  \item Report bugs at \url{https://github.com/thomasp85/fireproof/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Other contributors:
\itemize{
  \item Posit, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
