% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{get_path}
\alias{get_path}
\title{Extract the path from a URL}
\usage{
get_path(url, root = NULL)
}
\arguments{
\item{url}{The url to extract the path from}

\item{root}{An optional root to remove from the path as well}
}
\value{
The "path" part of the URL
}
\description{
This function is a simple helper that extract the path part of a URL. It is
useful when constructing OAuth 2.0 derived authenticators for the
\code{redirect_path} argument.
}
\examples{
get_path("https://example.com/auth")

get_path("https://example.com/api/auth", root = "/api")

}
\keyword{internal}
