% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_basic.R
\name{auth_basic}
\alias{auth_basic}
\title{Basic authentication plugin}
\usage{
auth_basic(
  authenticator,
  user_info = NULL,
  realm = "private",
  name = "BasicAuth"
)
}
\arguments{
\item{authenticator}{A function that will be called with the arguments
\code{username}, \code{password}, \code{realm}, \code{request}, and \code{response} and returns \code{TRUE}
if the user is valid, and \code{FALSE} otherwise. If the function returns a
character vector it is considered to be authenticated and the return value
will be understood as scopes the user is granted.}

\item{user_info}{A function to extract user information from the
username. It is called with two arguments: \code{user} and \code{setter},
the first being the username used for the successful authentication, the
second being a function that must be called in the end with the relevant
information (see \code{\link[=user_info]{user_info()}}).}

\item{realm}{The realm this authentication corresponds to. Will be returned
to the client on a failed authentication attempt to inform them of the
credentials required, though most often these days it is kept from the user.}

\item{name}{The name of the authentication}
}
\value{
An \link{AuthBasic} R6 object
}
\description{
Basic authentication is a HTTP scheme that sends username and password as a
\code{:} separated, base64 encoded string in the authorization header. Because it
is effectively send in plain text (base64 encoding can easily be decoded)
this should only ever be used along with other security measures such as
https/ssl to avoid username and passwords being snooped from the request.
}
\details{
This authenticator will use a user provided function to test a
username/password pair. It is up to the user to handle the storage and
testing of the passwords in a sensible and responsible way. See
\code{\link[sodium:password]{sodium::password_store()}} for a good first step towards responsible design.

If the authentication passes, the username from the authorization header is
written to the \code{username} data slot in the request
}
\section{User information}{
\code{auth_basic()} automatically adds \link[=user_info]{user information} after
authentication. By default it will set the \code{provider} field to \code{"local"} and
the \code{id} field to the username of used for logging in. Further, it will set
the \code{scopes} field to any scopes returned by the \code{authenticator} function.
}

\examples{
# Create an authenticator of dubious quality
basic <- auth_basic(
  authenticator = function(user, password) {
    user == "thomas" && password == "pedersen"
  },
  user_info = function(user, setter) {
    setter(
      name_given = "Thomas",
      name_middle = "Lin",
      name_family = "Pedersen"
    )
  }
)

# Add it to a fireproof plugin
fp <- Fireproof$new()
fp$add_auth(basic, "basic_auth")

# Use it in an endpoint
fp$add_auth_handler("get", "/*", basic_auth)

}
