% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_basic.R
\name{AuthBasic}
\alias{AuthBasic}
\title{R6 class for the Basic authentication scheme}
\description{
This class encapsulates the logic of the
\href{https://datatracker.ietf.org/doc/html/rfc7617}{Basic authentication scheme}.
See \code{\link[=auth_basic]{auth_basic()}} for more information.
}
\section{Super class}{
\code{\link[fireproof:Auth]{fireproof::Auth}} -> \code{AuthBasic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthBasic-new}{\code{AuthBasic$new()}}
\item \href{#method-AuthBasic-check_request}{\code{AuthBasic$check_request()}}
\item \href{#method-AuthBasic-reject_response}{\code{AuthBasic$reject_response()}}
\item \href{#method-AuthBasic-clone}{\code{AuthBasic$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="fireproof" data-topic="Auth" data-id="forbid_user"><a href='../../fireproof/html/Auth.html#method-Auth-forbid_user'><code>fireproof::Auth$forbid_user()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthBasic-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthBasic-new}{}}}
\subsection{Method \code{new()}}{
Constructor for the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthBasic$new(authenticator, realm = "private", name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{authenticator}}{A function that takes a username, password, realm,
request, and response and returns \code{TRUE} if the pair is valid, and \code{FALSE}
otherwise. If the function sets a character vector of scopes for the user in
the \code{auth_scope} field of the response data
(\code{response$set_data("auth_scope", ...)}) then it will be tested against the
scopes needed for the specific endpoint}

\item{\code{realm}}{The realm this authentication corresponds to. Will be returned
to the client on a failed authentication attempt to inform them of the
credentials required, though most often these days it is kept from the user.}

\item{\code{name}}{The name of the authentication}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthBasic-check_request"></a>}}
\if{latex}{\out{\hypertarget{method-AuthBasic-check_request}{}}}
\subsection{Method \code{check_request()}}{
A function that validates an incoming request, returning
\code{TRUE} if it is valid and \code{FALSE} if not. It decodes the credentials in
the \code{Authorization} header, splits it into username and password and then
calls the authenticator function provided at construction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthBasic$check_request(request, response, keys, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{request}}{The request to validate as a \link[reqres:Request]{Request}
object}

\item{\code{response}}{The corresponding response to the request as a
\link[reqres:Response]{Response} object}

\item{\code{keys}}{A named list of path parameters from the path matching}

\item{\code{...}}{Ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthBasic-reject_response"></a>}}
\if{latex}{\out{\hypertarget{method-AuthBasic-reject_response}{}}}
\subsection{Method \code{reject_response()}}{
Upon rejection this scheme sets the response status to \code{401}
and sets the \code{WWW-Authenticate} header to
\verb{Basic realm="<realm>", charset=UTF-8}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthBasic$reject_response(response, scope)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response}}{The response object}

\item{\code{scope}}{The scope of the endpoint}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthBasic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthBasic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthBasic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
