% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fireproof-package.R
\docType{package}
\name{fireproof-package}
\alias{fireproof}
\alias{fireproof-package}
\title{fireproof: Authentication and Authorization for 'fiery' servers}
\description{
Provides a set of plugins for 'fiery' that supports various forms of authorization and authentication.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/fireproof}
  \item Report bugs at \url{https://github.com/thomasp85/fireproof/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
